+++ src/main/java/com/invadermonky/magicultureintegrations/util/ModIds.java
@@ -7,10 +7,13 @@ import com.animania.Animania;
 import com.bewitchment.Bewitchment;
 import com.blakebr0.mysticalagriculture.MysticalAgriculture;
 import com.charles445.simpledifficulty.SimpleDifficulty;
+import com.github.alexthe666.rats.RatsMod;
+import com.lothrazar.cyclicmagic.util.Const;
 import com.oblivioussp.spartanweaponry.ModSpartanWeaponry;
 import com.pam.harvestcraft.Reference;
 import com.tmtravlr.qualitytools.QualityToolsMod;
 import crazypants.enderio.base.EnderIO;
+import de.ellpeck.actuallyadditions.mod.ActuallyAdditions;
 import de.ellpeck.naturesaura.NaturesAura;
 import epicsquid.mysticalworld.MysticalWorld;
 import epicsquid.roots.Roots;
@@ -19,6 +22,7 @@ import ic2.core.IC2;
 import josephcsible.oreberries.OreberriesMod;
 import knightminer.tcomplement.TinkersComplement;
 import mcjty.theoneprobe.TheOneProbe;
+import mod.icarus.crimsonrevelations.NewCrimsonRevelations;
 import net.blay09.mods.cookingforblockheads.CookingForBlockheads;
 import org.zeith.thaumicadditions.InfoTAR;
 import rustic.core.Rustic;
@@ -36,6 +40,7 @@ import wile.engineersdecor.ModEngineersDecor;
 import javax.annotation.Nullable;
 
 public enum ModIds {
+    actually_additions(ConstIds.actually_additions),
     agricraft(ConstIds.agricraft),
     animania(ConstIds.animania),
     astral_sorcery(ConstIds.astral_sorcery),
@@ -45,6 +50,7 @@ public enum ModIds {
     bloodmagic(ConstIds.bloodmagic),
     botania(ConstIds.botania),
     cooking_for_blockheads(ConstIds.cooking_for_blockheads),
+    cyclic(ConstIds.cyclic),
     embers(ConstIds.embers),
     enderio(ConstIds.enderio),
     engineers_decor(ConstIds.engineers_decor),
@@ -52,14 +58,17 @@ public enum ModIds {
     harvestcraft(ConstIds.harvestcraft),
     immersive_engineering(ConstIds.immersive_engineering),
     industrial_craft(ConstIds.industrialcraft),
+    industrial_foregoing(ConstIds.industrial_foregoing),
     mystical_agriculture(ConstIds.mystical_agriculture),
     mystical_world(ConstIds.mystical_world),
     natures_aura(ConstIds.natures_aura),
+    new_crimson_revelations(ConstIds.new_crimson_revelations),
     oreberries(ConstIds.oreberries),
     ore_stages(ConstIds.ore_stages),
     patchouli(ConstIds.patchouli),
     quality_tools(ConstIds.quality_tools),
     quark(ConstIds.quark),
+    rats(ConstIds.rats),
     redstone_paste(ConstIds.redstone_paste),
     roots(ConstIds.roots),
     rustic(ConstIds.rustic),
@@ -101,6 +110,7 @@ public enum ModIds {
 
 
     public static class ConstIds {
+        public static final String actually_additions = ActuallyAdditions.MODID;
         public static final String agricraft = "agricraft";
         public static final String animania = Animania.MODID;
         public static final String astral_sorcery = AstralSorcery.MODID;
@@ -110,6 +120,7 @@ public enum ModIds {
         public static final String bloodmagic = BloodMagic.MODID;
         public static final String botania = LibMisc.MOD_ID;
         public static final String cooking_for_blockheads = CookingForBlockheads.MOD_ID;
+        public static final String cyclic = Const.MODID;
         public static final String embers = Embers.MODID;
         public static final String enderio = EnderIO.MODID;
         public static final String engineers_decor = ModEngineersDecor.MODID;
@@ -117,14 +128,17 @@ public enum ModIds {
         public static final String harvestcraft = Reference.MODID;
         public static final String immersive_engineering = ImmersiveEngineering.MODID;
         public static final String industrialcraft = IC2.MODID;
+        public static final String industrial_foregoing = com.buuz135.industrial.utils.Reference.MOD_ID;
         public static final String mystical_agriculture = MysticalAgriculture.MOD_ID;
         public static final String mystical_world = MysticalWorld.MODID;
         public static final String natures_aura = NaturesAura.MOD_ID;
+        public static final String new_crimson_revelations = NewCrimsonRevelations.MODID;
         public static final String oreberries = OreberriesMod.MODID;
         public static final String ore_stages = "orestages";
         public static final String patchouli = Patchouli.MOD_ID;
         public static final String quality_tools = QualityToolsMod.MOD_ID;
         public static final String quark = vazkii.quark.base.lib.LibMisc.MOD_ID;
+        public static final String rats = RatsMod.MODID;
         public static final String redstone_paste = "redstonepaste";
         public static final String roots = Roots.MODID;
         public static final String rustic = Rustic.MODID;